add arg

+++ docs/build_docs/build.sh
@@ -1,32 +1,80 @@
+#i!/bin/bash
+
+help () {
+    echo ""
+    echo "Help:"
+    echo "$0 or $0 local"
+    echo "    Build html for local test, not merge to gh-pages branch"
+    echo "$0 version"
+    echo "    Build for version (version.py), then merge & push to gh-pages branch"
+    echo "$0 latest"
+    echo "    Build for latest code, then merge & push to gh-pages branch"
+}
 
 if [ ! -n "$1" ]; then
-  TAG=latest
+  ACT=only_build_local
 else
-  TAG=version
+  if [ "$1" == "version"  ]; then
+    ACT=build_version
+  elif [ "$1" == "latest"  ]; then
+    ACT=build_latest
+  elif [ "$1" == "local"  ]; then
+    ACT=only_build_local
+  elif [ "$1" == "help"  ]; then
+    help
+    exit 0
+  else
+    echo "Wrong parameter \"$1\""
+    help
+    exit 1
+  fi
 fi
 
-echo "Update ${TAG} folder"
+echo "ACT is ${ACT}"
 
-if [[ ${TAG} == "latest" ]]; then
+if [ ${ACT} == "only_build_local" ]; then
   UPDATE_LATEST_FOLDER=1
-  UPDATE_VERSION_FOLDER=0
-else
+  UPDATE_VERSION_FOLDER=1
+  CHECKOUT_GH_PAGES=0
+  PUSH_GH_PAGES=0
+
+elif [ ${ACT} == "build_version" ]; then
   UPDATE_LATEST_FOLDER=0
   UPDATE_VERSION_FOLDER=1
+  CHECKOUT_GH_PAGES=1
+  PUSH_GH_PAGES=1
+elif [ ${ACT} == "build_latest" ]; then
+  UPDATE_LATEST_FOLDER=1
+  UPDATE_VERSION_FOLDER=0
+  CHECKOUT_GH_PAGES=1
+  PUSH_GH_PAGES=1
 fi
 
-CHECKOUT_GH_PAGES=1
-PUSH_GH_PAGES=1
 WORK_DIR=../../build_tmp
-rm -rf ${WORK_DIR}
+if [ ! -d ${WORK_DIR} ]; then
+  echo "no ${WORK_DIR}"
+else
+  cp -rf ${WORK_DIR}/env_sphinx /tmp/
+  rm -rf ${WORK_DIR}
+fi
+
 mkdir -p ${WORK_DIR}
 cp -rf ./* ${WORK_DIR}
 
 cd ${WORK_DIR}
 
-if [ ! -d env_sphinx ]; then
-    bash pip_set_env.sh
+if [ ! -d /tmp/env_sphinx ]; then
+  echo "no /tmp/env_sphinx"
+else
+  echo "restore env_sphinx"
+  cp -r /tmp/env_sphinx ./
 fi
+
+if [ ! -d env_sphinx ]; then
+  echo "create env_sphinx"
+  bash pip_set_env.sh
+fi
+
 source env_sphinx/bin/activate
 
 cp -rf ../docs/ ./source
@@ -54,20 +102,10 @@ else
   exit 1
 fi
 
-if [[ ${CHECKOUT_GH_PAGES} -eq 1 ]]; then
-  git checkout -b gh-pages
-  git branch --set-upstream-to=origin/gh-pages gh-pages
-  git pull
-  git fetch origin
-  git reset --hard origin/gh-pages
-else
-  echo "skip pull gh-pages"
-fi
-
 
 VERSION=`cat source/version.txt`
-DST_FOLDER=../${VERSION}
-LATEST_FOLDER=../latest
+DST_FOLDER=./${VERSION}
+LATEST_FOLDER=./latest
 SRC_FOLDER=build/html
 
 if [[ ${UPDATE_VERSION_FOLDER} -eq 1 ]]; then
@@ -96,24 +134,48 @@ fi
 
 echo "Create document is done"
 
-echo "UPDATE_LATEST_FOLDER=${UPDATE_LATEST_FOLDER}"
-echo "UPDATE_VERSION_FOLDER=${UPDATE_VERSION_FOLDER}"
+if [[ ${CHECKOUT_GH_PAGES} -eq 1 ]]; then
+  git checkout -b gh-pages
+  git branch --set-upstream-to=origin/gh-pages gh-pages
+  git pull
+  git fetch origin
+  git reset --hard origin/gh-pages
 
-if [[ ${PUSH_GH_PAGES} -eq 1 ]]; then
   if [[ ${UPDATE_VERSION_FOLDER} -eq 1 ]]; then
-    echo "git add ${DST_FOLDER} ../versions.html"
-    git add ${DST_FOLDER} ../versions.html
+    python update_version.py ${DST_FOLDER} ${VERSION}
+    cp -rf ${DST_FOLDER} ../
   fi
 
   if [[ ${UPDATE_LATEST_FOLDER} -eq 1 ]]; then
-    echo "git add ${LATEST_FOLDER}"
-    git add ${LATEST_FOLDER}
+    python update_version.py ${LATEST_FOLDER} ${VERSION}
+    cp -rf ${LATEST_FOLDER} ../
+  fi
+
+else
+  echo "skip pull gh-pages"
+fi
+
+echo "UPDATE_LATEST_FOLDER=${UPDATE_LATEST_FOLDER}"
+echo "UPDATE_VERSION_FOLDER=${UPDATE_VERSION_FOLDER}"
+
+ROOT_DST_FOLDER=../${VERSION}
+ROOT_LATEST_FOLDER=../latest
+
+if [[ ${PUSH_GH_PAGES} -eq 1 ]]; then
+  if [[ ${UPDATE_VERSION_FOLDER} -eq 1 ]]; then
+    echo "git add ${ROOT_DST_FOLDER} ../versions.html"
+    git add ${ROOT_DST_FOLDER} ../versions.html
+  fi
+
+  if [[ ${UPDATE_LATEST_FOLDER} -eq 1 ]]; then
+    echo "git add ${ROOT_LATEST_FOLDER}"
+    git add ${ROOT_LATEST_FOLDER}
   fi
   git commit -m "update for ${VERSION}"
   git push origin gh-pages
   echo "git push origin gh-pages is done!"
 else
-  echo "Skip push"	
+  echo "Skip push"
 fi
 
 if [[ $? -eq 0 ]]; then
